refactor: rename user files
